Separating tests for reducers
